Add UUIDv6 support
