+++ Package.resolved
@@ -2,12 +2,21 @@
   "object": {
     "pins": [
       {
+        "package": "swift-asn1",
+        "repositoryURL": "https://github.com/apple/swift-asn1.git",
+        "state": {
+          "branch": null,
+          "revision": "7faebca1ea4f9aaf0cda1cef7c43aecd2311ddf6",
+          "version": "1.3.0"
+        }
+      },
+      {
         "package": "swift-crypto",
         "repositoryURL": "https://github.com/apple/swift-crypto.git",
         "state": {
           "branch": null,
-          "revision": "127d3745c37b5705e4bc8d16c7951c48dcc3332c",
-          "version": "2.0.0"
+          "revision": "ff0f781cf7c6a22d52957e50b104f5768b50c779",
+          "version": "3.10.0"
         }
       }
     ]
